+++ packages/x-components/src/x-modules/next-queries/store/index.ts
@@ -4,5 +4,6 @@ export * from './actions/set-query-from-last-history-query.action';
 export * from './emitters';
 export { request as nextQueriesRequest } from './getters/request.getter';
 export * from './getters/next-queries.getter';
+export * from './getters/next-queries-query.getter';
 export * from './module';
 export * from './types';